data/requirements.txt
# boto3==1.26.108
requests==2.28.2
azure-storage-blob==12.15.0
# tqdm==4.65.0